Backend/requirements.txt
alembic==1.18.3
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
ordered-set==4.1.0
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
PyJWT==2.11.0
python-dotenv==1.2.1
python-jose==3.5.0
python-slugify==8.0.4
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.1.1
